rf: move subject filtering, processing groups prior to config file save
